PoC of repository in repository injection
