+++ src/test/java/org/trd/app/teknichrono/model/jpa/TestBeaconRepository.java
@@ -1,11 +1,11 @@
-package org.trd.app.teknichrono.service;
+package org.trd.app.teknichrono.model.jpa;
 
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.trd.app.teknichrono.model.dto.BeaconDTO;
 import org.trd.app.teknichrono.model.jpa.Beacon;
@@ -19,19 +19,19 @@ import org.trd.app.teknichrono.util.exception.NotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.Mockito.atLeastOnce;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-public class TestBeaconService {
-
+public class TestBeaconRepository {
 
   private long id = 1L;
 
@@ -41,12 +41,9 @@ public class TestBeaconService {
   @Mock
   private PingRepository pingRepository;
 
-  @Mock
-  private BeaconRepository beaconRepository;
-
+  @Spy
   @InjectMocks
-  private BeaconService service;
-
+  private BeaconRepository beaconRepository;
 
   public Beacon newBeacon(long number, long pilotId) {
     Beacon beacon = new Beacon();
@@ -71,52 +68,54 @@ public class TestBeaconService {
   @Test
   public void createsBeaconWithCompletePilotIfIdProvided() throws NotFoundException {
     Beacon entity = newBeacon(999, 9);
-    when(pilotRepository.findById(9L)).thenReturn(entity.getPilot());
-    service.create(entity);
+    Pilot pilot = entity.getPilot();
+
+    when(pilotRepository.findById(9L)).thenReturn(pilot);
+    doNothing().when(beaconRepository).persist(entity);
+
+    beaconRepository.create(entity);
+
     verify(pilotRepository).findById(9L);
-
-    ArgumentCaptor<Beacon> captor = ArgumentCaptor.forClass(Beacon.class);
-    verify(beaconRepository).persist(captor.capture());
-
-    assertThat(captor.getValue().getId()).isEqualTo(entity.getId());
-    assertThat(captor.getValue().getNumber()).isEqualTo(entity.getNumber());
+    verify(beaconRepository).persist(entity);
   }
 
   @Test
   public void deleteByIdRemovesBeaconFromPilot() throws NotFoundException {
     Beacon entity = newBeacon(999, 9);
-    when(beaconRepository.findById(anyLong())).thenReturn(entity);
-    service.deleteById(entity.id);
-    ArgumentCaptor<Beacon> bCaptor = ArgumentCaptor.forClass(Beacon.class);
-    verify(beaconRepository).delete(bCaptor.capture());
-    assertThat(bCaptor.getValue().getId()).isEqualTo(entity.id);
+    Pilot pilot = entity.getPilot();
 
-    ArgumentCaptor<Pilot> captor = ArgumentCaptor.forClass(Pilot.class);
-    verify(pilotRepository, atLeastOnce()).persist(captor.capture());
-    List<Pilot> pilots = captor.getAllValues();
-    for (Pilot p : pilots) {
-      assertThat(p.getCurrentBeacon()).isNull();
-    }
+    doReturn(entity).when(beaconRepository).findById(anyLong());
+    doNothing().when(beaconRepository).delete(entity);
+
+    beaconRepository.deleteById(entity.id);
+
+    verify(beaconRepository).delete(entity);
+    verify(pilotRepository).persist(pilot);
+    assertThat(pilot.getCurrentBeacon()).isNull();
   }
 
   @Test
   public void deleteByIdRemovesBeaconFromPings() throws NotFoundException {
     Beacon entity = newBeacon(999, 9);
-    when(beaconRepository.findById(anyLong())).thenReturn(entity);
-    service.deleteById(entity.getId());
-    ArgumentCaptor<Ping> captor = ArgumentCaptor.forClass(Ping.class);
-    verify(pingRepository, atLeastOnce()).persist(captor.capture());
-    List<Ping> pings = captor.getAllValues();
-    assertThat(pings.size()).isEqualTo(entity.getPings().size());
-    for (Ping p : pings) {
-      assertThat(p.getBeacon()).isNull();
+    List<Ping> pings = entity.getPings();
+
+    doReturn(entity).when(beaconRepository).findById(anyLong());
+    doNothing().when(beaconRepository).delete(entity);
+
+    beaconRepository.deleteById(entity.id);
+
+    verify(beaconRepository).delete(entity);
+
+    for (Ping ping : pings) {
+      verify(pingRepository).persist(ping);
+      assertThat(ping.getBeacon()).isNull();
     }
   }
 
   @Test
   public void deleteByIdReturnsErrorIfBeaconDoesNotExist() {
-    Beacon entity = newBeacon(999, 9);
-    assertThrows(NotFoundException.class, () -> service.deleteById(entity.id));
+    doReturn(null).when(beaconRepository).findById(42L);
+    assertThrows(NotFoundException.class, () -> beaconRepository.deleteById(42L));
     verify(beaconRepository, never()).delete(any());
   }
 
@@ -132,11 +131,11 @@ public class TestBeaconService {
     entities.add(entity3);
 
     PanacheQuery<Beacon> query = mock(PanacheQuery.class);
-    when(beaconRepository.findAll()).thenReturn(query);
+    doReturn(query).when(beaconRepository).findAll();
     when(query.page(any())).thenReturn(query);
     when(query.stream()).thenReturn(entities.stream());
 
-    List<BeaconDTO> beacons = service.findAll(null, null);
+    List<BeaconDTO> beacons = beaconRepository.findAll(null, null);
     assertThat(beacons).isNotNull();
     assertThat(beacons).hasSize(3);
 
@@ -153,11 +152,11 @@ public class TestBeaconService {
     entities.add(entity1);
 
     PanacheQuery<Beacon> query = mock(PanacheQuery.class);
-    when(beaconRepository.findAll()).thenReturn(query);
+    doReturn(query).when(beaconRepository).findAll();
     when(query.page(any())).thenReturn(query);
     when(query.stream()).thenReturn(entities.stream());
 
-    List<BeaconDTO> beacons = service.findAll(1, 1);
+    List<BeaconDTO> beacons = beaconRepository.findAll(1, 1);
     assertThat(beacons).isNotNull();
     assertThat(beacons).hasSize(1);
   }
